+++ docs/guide/changelog.md
@@ -6,6 +6,7 @@
 ## 1.6.1 <Tag color="volcano" size="small">2024-08-26</Tag>
 
 - 优化并更新 [输入框 Input](https://themusecatcher.github.io/vue-amazing-ui/guide/components/input.html) 组件
+- 优化并更新 [气泡卡片 Popover](https://themusecatcher.github.io/vue-amazing-ui/guide/components/popover.html) 组件，新增 `titleStyle` `contentStyle` 属性
 - 组件库代码优化
 
 ## 1.6.0 <Tag color="volcano" size="small">2024-08-23</Tag>